+++ rtic/src/export/riscv_esp32c3.rs
@@ -1,7 +1,6 @@
-use esp32c3::INTERRUPT_CORE0; //priority threshold control
+use esp32c3::INTERRUPT_CORE0;
 pub use esp32c3::{Interrupt, Peripherals};
-pub use riscv::interrupt;
-pub use riscv::register::mcause; //low level interrupt enable/disable
+pub use riscv::{interrupt, register::mcause};
 
 #[cfg(all(feature = "riscv-esp32c3", not(feature = "riscv-esp32c3-backend")))]
 compile_error!("Building for the esp32c3, but 'riscv-esp32c3-backend not selected'");
@@ -138,28 +137,20 @@ pub fn unpend(int: Interrupt) {
 }
 
 pub fn enable(int: Interrupt, prio: u8, cpu_int_id: u8) {
-    const INTERRUPT_MAP_BASE: u32 = 0x600c2000; //this isn't exposed properly in the PAC,
-                                                //should maybe figure out a workaround that
-                                                //doesnt involve raw pointers.
-                                                //Again, this is how they do it in the HAL
-                                                //but i'm really not a fan.
-    let interrupt_number = int as isize;
-    let cpu_interrupt_number = cpu_int_id as isize;
-
     unsafe {
-        let intr_map_base = INTERRUPT_MAP_BASE as *mut u32;
-        intr_map_base
-            .offset(interrupt_number)
-            .write_volatile(cpu_interrupt_number as u32);
-        //map peripheral interrupt to CPU interrupt
+        // Map the peripheral interrupt to a CPU interrupt:
+        (INTERRUPT_CORE0::ptr() as *mut u32)
+            .offset(int as isize)
+            .write_volatile(cpu_int_id as u32);
+
+        // Set the interrupt's priority:
+        (*INTERRUPT_CORE0::ptr())
+            .cpu_int_pri(cpu_int_id as usize)
+            .modify(|_, w| w.bits(prio as u32));
+
+        // Finally, enable the CPU interrupt:
         (*INTERRUPT_CORE0::ptr())
             .cpu_int_enable()
-            .modify(|r, w| w.bits((1 << cpu_interrupt_number) | r.bits())); //enable the CPU interupt.
-        let intr = INTERRUPT_CORE0::ptr();
-        let intr_prio_base = (*intr).cpu_int_pri_0().as_ptr();
-
-        intr_prio_base
-            .offset(cpu_interrupt_number)
-            .write_volatile(prio as u32);
+            .modify(|r, w| w.bits((1 << cpu_int_id) | r.bits()));
     }
 }